KAM-2072: moving Error screens to components
