+++ apps/consultation-portal/components/Error/index.ts
@@ -1,0 +1,3 @@
+export { default as Error404 } from './Error404/Error404'
+export { default as Error500 } from './Error500/Error500'
+export { default as CustomError } from './CustomError/CustomError'
