+++ apps/consultation-portal/components/index.ts
@@ -13,3 +13,4 @@ export { default as Pagination } from './Pagination/Pagination'
 export { default as SearchAndSort } from './SearchAndSort/SearchAndSort'
 export { default as SearchAndSortPartial } from './SearchAndSort/SearchAndSortPartialData'
 export { default as LogoText } from './LogoText/LogoText'
+export * from './Error'
